Create circle_actuation.c
Code prints out x and y actuation lengths for making the nozzle go in a circle (nozzle is always 10 degrees from pointing straight). This code assumes that the zero position of the actuators is half of its full extension. It also assumes that the distance from the center of the gimble to the contact point of the actuator is 5 inches. There are 100 increments around the hypothetical circle.